+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Changed
 - Updated versions of the dependencies: `pylangacq >= 0.17.0` and `wordseg >= 0.0.4`.
 - Restructured the repository to use top-level `src/` and `tests/` directories.
+- Replaced tagger.pickle with tagger.json.
 
 ### Deprecated
 ### Removed